fix: split claude_json cleanup tests into own file, cover error paths for 100% line floor
