+++ src/routines/cleanup/cleanup_tests.rs
@@ -553,161 +553,6 @@ fn kill_sessions_for_deleted_routine_kills_the_live_workbench_session() {
         "moadim-cleanup-kill-deleted-{}",
         uuid::Uuid::new_v4()
     ));
-    let _ = std::fs::remove_dir_all(&base);
-    std::fs::create_dir_all(&base).unwrap();
-    touch_dir(&base, "expired-100");
-    std::fs::write(base.join("expired-100").join("inner"), b"x").unwrap();
-
-    // Strip write permission from the parent so removing the child directory fails.
-    let mut perms = std::fs::metadata(&base).unwrap().permissions();
-    perms.set_mode(0o555);
-    std::fs::set_permissions(&base, perms).unwrap();
-
-    let now = 1000;
-    let ttl_for = |_slug: &str| 500u64; // age 900 > 500 -> expired
-    let dead = |_session: &str| false;
-    let removed = reap_dir(
-        &base,
-        now,
-        &ttl_for,
-        &never_expires_runtime,
-        &dead,
-        &noop_kill,
-    );
-    // A read-only parent makes `remove_dir_all` fail for an unprivileged user, so
-    // the directory survives and the Err arm runs (0 removed). Root bypasses the
-    // permission check; tolerate that by only asserting consistency.
-    if base.join("expired-100").exists() {
-        assert_eq!(removed, 0);
-    }
-
-    // Restore permissions so cleanup can proceed.
-    let mut perms = std::fs::metadata(&base).unwrap().permissions();
-    perms.set_mode(0o755);
-    std::fs::set_permissions(&base, perms).unwrap();
-    std::fs::remove_dir_all(&base).unwrap();
-}
-
-#[test]
-fn cleanup_expired_workbenches_scans_real_workbenches_dir() {
-    // Drives the public entry point so `cleanup_expired_workbenches` resolves the real
-    // `workbenches_dir()` (honouring MOADIM_HOME_OVERRIDE) and `tmux_session_alive` runs as the
-    // injected liveness check. With an empty store every slug falls back to MAX_TTL_SECS, so we
-    // stamp the expired workbench far enough in the past to exceed that cap.
-    let home = std::env::temp_dir().join(format!("moadim-cleanup-{}", uuid::Uuid::new_v4()));
-    let previous = std::env::var_os("MOADIM_HOME_OVERRIDE");
-    // SAFETY: tests in this crate run single-threaded (RUST_TEST_THREADS=1); restored below.
-    unsafe {
-        std::env::set_var("MOADIM_HOME_OVERRIDE", &home);
-    }
-
-    let workbenches = crate::paths::workbenches_dir();
-    std::fs::create_dir_all(&workbenches).unwrap();
-    // An expired (timestamp 1) finished workbench whose tmux session is absent -> reaped.
-    std::fs::create_dir_all(workbenches.join("orphan-1")).unwrap();
-    // A workbench triggered "now-ish" so it is not yet expired -> kept.
-    let fresh_ts = now_secs();
-    std::fs::create_dir_all(workbenches.join(format!("recent-{fresh_ts}"))).unwrap();
-    // A non-workbench directory (no numeric suffix) -> skipped.
-    std::fs::create_dir_all(workbenches.join("notawb")).unwrap();
-
-    let store = super::super::model::new_store();
-    let removed = cleanup_expired_workbenches(&store);
-
-    // The orphaned, expired, session-less workbench is removed; the others survive.
-    assert!(removed >= 1, "expected at least the orphan to be reaped");
-    assert!(!workbenches.join("orphan-1").exists());
-    assert!(workbenches.join(format!("recent-{fresh_ts}")).exists());
-    assert!(workbenches.join("notawb").exists());
-
-    // SAFETY: single-threaded harness; restore the saved override.
-    unsafe {
-        match previous {
-            Some(value) => std::env::set_var("MOADIM_HOME_OVERRIDE", value),
-            None => std::env::remove_var("MOADIM_HOME_OVERRIDE"),
-        }
-    }
-    let _ = std::fs::remove_dir_all(&home);
-}
-
-#[test]
-fn cleanup_expired_workbenches_prunes_the_reaped_workbench_claude_json_entry() {
-    // Reaping a workbench must also prune its `projects[<workbench>]` entry from `~/.claude.json`
-    // (issue #430), so that shared config does not accumulate one dead entry per reaped run. Drives
-    // the public entry point end-to-end: a real `~/.claude.json` under the overridden home, seeded
-    // with an entry for the workbench that is about to be reaped plus an unrelated entry that must
-    // survive.
-    let home = std::env::temp_dir().join(format!(
-        "moadim-cleanup-claude-json-{}",
-        uuid::Uuid::new_v4()
-    ));
-    let previous = std::env::var_os("MOADIM_HOME_OVERRIDE");
-    // SAFETY: tests in this crate run single-threaded (RUST_TEST_THREADS=1); restored below.
-    unsafe {
-        std::env::set_var("MOADIM_HOME_OVERRIDE", &home);
-    }
-
-    let workbenches = crate::paths::workbenches_dir();
-    std::fs::create_dir_all(&workbenches).unwrap();
-    std::fs::create_dir_all(workbenches.join("orphan-1")).unwrap();
-    let reaped_workbench = workbenches.join("orphan-1");
-    let reaped_key = reaped_workbench.to_string_lossy().into_owned();
-
-    let claude_json = crate::paths::claude_json_path().unwrap();
-    std::fs::write(
-        &claude_json,
-        format!(
-            r#"{{"projects":{{"{reaped_key}":{{"hasTrustDialogAccepted":true}},"/kept/wb":{{}}}}}}"#
-        ),
-    )
-    .unwrap();
-
-    let store = super::super::model::new_store();
-    let removed = cleanup_expired_workbenches(&store);
-
-    assert!(removed >= 1, "expected the orphan to be reaped");
-    assert!(!reaped_workbench.exists());
-    let rewritten: serde_json::Value =
-        serde_json::from_str(&std::fs::read_to_string(&claude_json).unwrap()).unwrap();
-    let projects = rewritten.get("projects").unwrap().as_object().unwrap();
-    assert!(
-        !projects.contains_key(&reaped_key),
-        "the reaped workbench's projects entry must be pruned"
-    );
-    assert!(
-        projects.contains_key("/kept/wb"),
-        "unrelated projects entries must survive"
-    );
-
-    // SAFETY: single-threaded harness; restore the saved override.
-    unsafe {
-        match previous {
-            Some(value) => std::env::set_var("MOADIM_HOME_OVERRIDE", value),
-            None => std::env::remove_var("MOADIM_HOME_OVERRIDE"),
-        }
-    }
-    let _ = std::fs::remove_dir_all(&home);
-}
-
-#[cfg(unix)]
-#[test]
-fn cleanup_expired_workbenches_kills_a_live_hung_session() {
-    use std::os::unix::fs::PermissionsExt as _;
-
-    // Drives the public entry point against a *live* session so the watchdog path runs end-to-end:
-    // a stub `tmux` that always exits 0 makes `tmux_session_alive` report the session as running
-    // (exercising its `status.success()` mapping over a real process), which in turn makes
-    // `cleanup_expired_workbenches` consult its `max_runtime_for` bound. An ancient timestamp puts
-    // the run past the (empty-store default) max runtime, so the session is force-killed, the kill
-    // is noted in agent.log, and the workbench is reaped. Complements
-    // `cleanup_expired_workbenches_scans_real_workbenches_dir`, which covers the no-tmux path.
-    let home = std::env::temp_dir().join(format!("moadim-cleanup-hung-{}", uuid::Uuid::new_v4()));
-    std::fs::create_dir_all(&home).unwrap();
-    // A stub tmux that ignores its args and always succeeds, so has-session/kill-session both "work".
-    let stub_tmux = home.join("stub-tmux");
-    std::fs::write(&stub_tmux, b"#!/bin/sh\nexit 0\n").unwrap();
-    std::fs::set_permissions(&stub_tmux, std::fs::Permissions::from_mode(0o755)).unwrap();
-
     let prev_home = std::env::var_os("MOADIM_HOME_OVERRIDE");
     let prev_tmux = std::env::var_os("MOADIM_TMUX_BIN");
     // SAFETY: tests in this crate run single-threaded (RUST_TEST_THREADS=1); restored below.
